requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
amqtt==0.10.0a3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
awsebcli==3.19.3
boto3==1.17.8
botocore==1.20.8
cement==2.8.2
certifi==2019.3.9
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.1
dj-database-url==0.5.0
Django==3.1.5
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.2
docopt==0.6.2
EB==0.1.5
Flask==1.0.2
Flask-WTF==0.14.2
future==0.16.0
gunicorn==20.0.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
MarkupSafe==1.1.1
msgpack==1.0.2
multidict==5.1.0
passlib==1.7.4
pathspec==0.5.9
Pillow==6.0.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.1
PyYAML==5.4.1
requests==2.24.0
s3transfer==0.3.4
selenium==3.141.0
semantic-version==2.5.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.0
termcolor==1.1.0
transitions==0.8.8
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==16.4.3
virtualenvwrapper-win==1.2.5
wcwidth==0.1.9
websockets==8.1
Werkzeug==0.15.2
whitenoise==5.2.0
WTForms==2.2.1
yarl==1.6.3
zope.interface==5.3.0
